requirements.txt
django==3.0.5
djangorestframework==3.11.0
beautifulsoup4==4.9.0
requests==2.23.0
redis==3.4.1
psycopg2==2.8.5
selenium==3.141.0
celery==4.4.2
gunicorn==20.0.4
django-clear-cache==0.3
selenium-wire==1.0.11
django-prettyjson==0.4.1
django-jsonfield==1.4.0
python-telegram-bot==12.7
django-environ==0.4.5
